backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv==1.0.0

# OpenAI & LangChain
openai>=1.12.0  # Updated for httpx 0.26.0 compatibility (fixes 'proxies' parameter issue)
langchain>=0.3.0  # Updated to be compatible with langchain-openai>=0.1.0 and openai>=2.0.0
langchain-openai>=0.1.0  # Updated to support openai>=2.0.0
langgraph>=1.0.0  # Updated to be compatible with langchain>=0.3.0
langsmith>=0.1.17  # Updated to be compatible with langchain>=0.3.0 (requires langsmith>=0.1.17)

# Database
sqlalchemy==2.0.25
psycopg2-binary>=2.9.0
alembic==1.13.1

# Redis
redis==5.0.1
# hiredis==2.3.2  # Optional: Performance optimization, requires compilation on Windows

# HTTP Clients
httpx>=0.27.0  # Updated to be compatible with mcp>=1.0.0 (requires httpx>=0.27)
# aiohttp==3.9.1  # Optional: Requires C++ compiler on Windows. httpx is sufficient.

# Task Queue (Optional)
celery==5.3.4
flower==2.0.1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
# httpx already listed above in HTTP Clients section

# MCP Integration Dependencies
jira==3.6.0
# pyzoom==0.1.10  # Optional: Conflicts with pydantic>=2.0. Use Zoom API directly via httpx instead
pysnow==0.7.17
mcp>=1.0.0  # Model Context Protocol for tool standardization

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

